backend/segmenter/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pillow==11.0.0
numpy==2.1.3
transformers>=5.5.0,<5.6.0
accelerate>=1.0.0,<2.0.0
huggingface-hub>=0.30.0,<1.0.0
safetensors>=0.5.0,<1.0.0
